### 2017-07-26 #336 @ricleal
Arfon,

This software is part of our scientific pipeline.
It is used to submit asynchronously to a cluster (using ssh) python scripts. In our case, we have a user friendly web interface and all the complex fitting routines are submitted to this cluster. Once the fitting routine is finished the results are transparently sent back to the user.
Let me know if this is enough to clarify the research application of this software.
Thank you
