+++ django_remote_submission/local.py
@@ -1,0 +1,260 @@
+"""Provides a wrapper to mimick the remote but runs in local.
+
+This module is meant to be a general wrapper so that a ``LocalWrapper`` can
+also be created to run tests in continuous integration services where SSH is
+not available.
+
+"""
+
+from __future__ import absolute_import, unicode_literals, print_function
+import logging
+import os
+import textwrap
+import datetime
+import uuid
+
+try:
+    from shlex import quote as cmd_quote
+except ImportError:
+    from pipes import quote as cmd_quote
+
+from paramiko import (
+    AuthenticationException, BadHostKeyException,
+)
+from paramiko.client import SSHClient, AutoAddPolicy
+from django.utils.timezone import now
+import six
+
+logger = logging.getLogger(__name__)
+
+
+class LocalWrapper(object):
+    """
+
+    """
+
+    def __init__(self, hostname, username, port=None):
+        """Initialize the wrapper.
+        """
+        pass
+
+    def connect(self, password=None, public_key_filename=None):
+        """
+        """
+        pass
+
+    def close(self):
+        pass
+
+    def chdir(self, local_directory):
+        """
+        """
+        os.chdir(local_directory)
+
+    def open(self, filename, mode):
+        pass
+
+    def listdir_attr(self):
+        """Retrieve a list of files and their attributes.
+
+        Each object is guaranteed to have a ``filename`` attribute as well as
+        an ``st_mtime`` attribute, which gives the last modified time in
+        seconds.
+
+        """
+        class T(object):
+            def __init__(self, filename, st_mtime):
+                self.filename = filename
+                self.st_mtime = st_mtime
+            
+        filenames = os.listdir('.')
+        res = [T(filename, os.stat(filename).st_mtime)
+               for filename in filenames]
+        return res
+
+
+    def exec_command(self, args, workdir, timeout=None, stdout_handler=None,
+                     stderr_handler=None):
+        """Execute a command on the remote server.
+
+        An example of how to use this function::
+
+            from datetime import timedelta
+            wrapper.exec_command(
+                args=["ls", "-la", "."],
+                workdir="/",
+                timeout=timedelta(minute=5),
+                stdout_handler=lambda now, output: print('stdout, now, output),
+                stderr_handler=lambda now, output: print('stderr, now, output),
+            )
+
+        :param list(str) args: the command and arguments to run
+        :param str workdir: the directory to run the commands from
+        :param datetime.timedelta timeout: the timeout to use for the command
+        :param stdout_handler: a function that accepts ``now`` and ``output``
+            parameters and is called when new output appears on stdout.
+        :param stderr_handler: a function that accepts ``now`` and ``output``
+            parameters and is called when new output appears on stderr.
+
+        """
+        chdir = self._make_command(['cd', workdir], None)
+        run = self._make_command(args, timeout)
+        command = '{} && {}'.format(chdir, run)
+        print('exec_command(command={!r})'.format(command))
+
+        transport = self._client.get_transport()
+        channel = transport.open_session()
+
+        channel.exec_command(command)
+        while True:
+            current_time = now()
+            if channel.recv_ready():
+                output = channel.recv(1024).decode('utf-8')
+                if stdout_handler is not None:
+                    stdout_handler(current_time, output)
+
+            if channel.recv_stderr_ready():
+                output = channel.recv_stderr(1024).decode('utf-8')
+                if stderr_handler is not None:
+                    stderr_handler(current_time, output)
+
+            if channel.exit_status_ready():
+                if channel.recv_ready() or channel.recv_stderr_ready():
+                    continue
+
+                if channel.recv_exit_status() == 0:
+                    return True
+                else:
+                    return False
[…]
