### 2017-08-10 ce8ad83
Corrected task log to be in celert
