+++ django_remote_submission/migrations/0001_initial.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Generated by Django 1.11.3 on 2017-07-06 21:05
+# Generated by Django 1.11.3 on 2017-08-09 19:01
 from __future__ import unicode_literals
 
 from django.conf import settings
@@ -31,8 +31,8 @@ class Migration(migrations.Migration):
                 ('arguments', django_remote_submission.models.ListField(help_text='The arguments used when running the interpreter', max_length=256, verbose_name='Command Arguments')),
             ],
             options={
-                'verbose_name_plural': 'interpreters',
                 'verbose_name': 'interpreter',
+                'verbose_name_plural': 'interpreters',
             },
         ),
         migrations.CreateModel(
@@ -51,8 +51,8 @@ class Migration(migrations.Migration):
                 ('owner', models.ForeignKey(help_text='The user that owns this job', on_delete=django.db.models.deletion.PROTECT, related_name='jobs', to=settings.AUTH_USER_MODEL, verbose_name='Job Owner')),
             ],
             options={
-                'verbose_name_plural': 'jobs',
                 'verbose_name': 'job',
+                'verbose_name_plural': 'jobs',
             },
         ),
         migrations.CreateModel(
@@ -65,8 +65,8 @@ class Migration(migrations.Migration):
                 ('job', models.ForeignKey(help_text='The job this log came from', on_delete=django.db.models.deletion.CASCADE, related_name='logs', to='django_remote_submission.Job', verbose_name='Log Job')),
             ],
             options={
-                'verbose_name_plural': 'logs',
                 'verbose_name': 'log',
+                'verbose_name_plural': 'logs',
             },
         ),
         migrations.CreateModel(
@@ -80,8 +80,8 @@ class Migration(migrations.Migration):
                 ('job', models.ForeignKey(help_text='The job this result came from', on_delete=django.db.models.deletion.CASCADE, related_name='results', to='django_remote_submission.Job', verbose_name='Result Job')),
             ],
             options={
-                'verbose_name_plural': 'results',
                 'verbose_name': 'result',
+                'verbose_name_plural': 'results',
             },
         ),
         migrations.CreateModel(
@@ -96,8 +96,8 @@ class Migration(migrations.Migration):
                 ('interpreters', models.ManyToManyField(to='django_remote_submission.Interpreter', verbose_name='List of interpreters available for this Server')),
             ],
             options={
-                'verbose_name_plural': 'servers',
                 'verbose_name': 'server',
+                'verbose_name_plural': 'servers',
             },
         ),
         migrations.AddField(
