### 2017-08-12 #336 @arfon
Thanks @ricleal. Since you submitted this package we've expanded a little on the requirements for the `paper.md` file. Could you take a look at the [updated guidelines](http://joss.theoj.org/about#paper_structure) and expand a little on what you have?

In the meantime, I'll look for a reviewer.
