### 2017-08-14 #366 @whedon
[REVIEW]: Django Remote Submission
**Submitting author:** @ricleal (<a href="http://orcid.org/0000-0002-9931-8304">Ricardo Miguel Ferraz Leal</a>)
**Repository:** <a href="https://github.com/ornl-ndav/django-remote-submission" target ="_blank">https://github.com/ornl-ndav/django-remote-submission</a>
**Version:** v0.11.1
