### 2017-08-14 #336 @arfon
@djmitche please head over to #366 for the actual review will happen.
