### 2017-08-15 ornl-ndav/django-remote-submission#33 @djmitche
No github tags for recent versions
I see releases up to 0.11.1 on pypi, but the tags on this repo don't include versions up that high.  Tags are helpful in trying to figure out what has changed when upgrading.
