### 2017-08-15 #366 @djmitche
*  Version: https://github.com/ornl-ndav/django-remote-submission/issues/33
* References:  I think this is OK, but there are no DOIs.  All of the referenced software (Django, Redis, Celery) probably doesn't have a DOI. @arfon am I missing anything there?

(I will continue later)
