### 2017-08-15 #366 @ricleal
Thanks @arfon and @djmitche for putting effort into reviewing this.
I was away the last few days but will have a look at the updated guidelines and expand a little on how this peace of software enables and addresses research challenges.
Thanks again
