### 2017-08-15 #366 @djmitche
While you're thinking about that, I'm also curious what value this adds over Celery itself.  The code may demonstrate that to me when I dig in, so I apologize if the question seems ignorant.
