### 2017-08-15 #366 @ricleal
Not ignorant at all. Celery is just an asynchronous job queue. We have build around that. Django uses the Database to provide data to views and hence to the browser. Our Jobs besides running our scientific routines must be constantly writing to the DB (e.g. changing the status of the job - submitted, failed, etc - as well writing the routine stdout/err, etc) so the browser can track the job and show as much information as possible.
A Job includes thus the ssh/sftp to the cluster, the database updates and the launching of the scientific routine itself. This Job is what we submit to Celery.
Hope that was clear.
