+++ paper.bib
@@ -36,4 +36,32 @@
   url    = {https://github.com/django/channels},
 }
 
+@Article{Mason2006,
+  Title                    = {The Spallation Neutron Source in Oak Ridge: A powerful tool for materials research},
+  Author                   = {T.E. Mason and D. Abernathy and I. Anderson and J. Ankner and T. Egami and G. Ehlers and A. Ekkebus and G. Granroth and M. Hagen and K. Herwig and J. Hodges and C. Hoffmann and C. Horak and L. Horton and F. Klose and J. Larese and A. Mesecar and D. Myles and J. Neuefeind and M. Ohl and C. Tulk and X-L. Wang and J. Zhao},
+  Journal                  = {Physica B: Condensed Matter},
+  Year                     = {2006},
+  Pages                    = {955 - 960},
+  Volume                   = {385},
+  Doi                      = {http://dx.doi.org/10.1016/j.physb.2006.05.281},
+  ISSN                     = {0921-4526},
+  Keywords                 = {Pulsed neutron},
+  Url                      = {http://www.sciencedirect.com/science/article/pii/S092145260601177X}
+}
+
+@article{doucet2017,
+  author = {M. Doucet and R. M. Ferraz Leal and T. C. Hobson},
+  title  = {ORNL Reflectivity Fitting Interface},
+  url    = {https://github.com/neutrons/web_reflectivity},
+  DOI={10.5281/zenodo.260178},
+  year={2017}, month={Jan}
+}
+
+
+@Online{Kienzle,
+  title={Refl1D: Interactive depth profile modeler},
+  author={Kienzle, PA and Krycka, JA and Patel, N},
+  url={https://github.com/reflectometry/refl1d}
+}
+
 @Comment{jabref-meta: databaseType:biblatex;}
