+++ AUTHORS.rst
@@ -6,6 +6,8 @@ Development Lead
 ----------------
 
 * Tanner Hobson <thobson125@gmail.com>
+* Mathieu Doucet <doucetm@ornl.gov>
+* Ricardo M. Ferraz Leal <ferrazlealrm@ornl.gov>
 
 Contributors
 ------------
