+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 -------
 
+0.11.2 (2017-08-15)
++++++++++++++++++++
+
+* Publication ready
+
 0.2.0 (2016-11-17)
 ++++++++++++++++++
 
