### 2017-08-15 f6fc4bc
Added info about the example and .env
