### 2017-08-15 4ab14e6
version bump and format bib
