+++ paper.bib
@@ -1,39 +1,46 @@
+% This file was created with JabRef 2.10.
 % Encoding: UTF-8
 
-@Online{Django,
-  author = {Django},
-  title  = {Django Web Application Framework},
-  url    = {https://www.djangoproject.com/},
-}
-
-@Online{PyPi,
-  author = {PyPi},
-  title  = {PyPI - the Python Package Index},
-  url    = {https://pypi.python.org/pypi/django-remote-submission},
-}
-
-@Online{django-remote-submission,
-  author = {django-remote-submission},
-  title  = {Django Remote Submission},
-  url    = {https://github.com/ornl-ndav/django-remote-submission},
-}
 
 @Online{Celery,
-  author = {Celery},
-  title  = {Celery: Distributed Task Queue},
-  url    = {http://www.celeryproject.org},
-}
-
-@Online{Redis,
-  author = {Redis},
-  title  = {Redis},
-  url    = {https://redis.io},
+  Title                    = {Celery: Distributed Task Queue},
+  Author                   = {Celery},
+  Url                      = {http://www.celeryproject.org}
 }
 
 @Online{Channels,
-  author = {Channels},
-  title  = {Django Channels},
-  url    = {https://github.com/django/channels},
+  Title                    = {Django Channels},
+  Author                   = {Channels},
+  Url                      = {https://github.com/django/channels}
+}
+
+@Online{Django,
+  Title                    = {Django Web Application Framework},
+  Author                   = {Django},
+  Url                      = {https://www.djangoproject.com/}
+}
+
+@Online{django-remote-submission,
+  Title                    = {Django Remote Submission},
+  Author                   = {django-remote-submission},
+  Url                      = {https://github.com/ornl-ndav/django-remote-submission}
+}
+
+@Article{doucet2017,
+  Title                    = {ORNL Reflectivity Fitting Interface},
+  Author                   = {M. Doucet and R. M. Ferraz Leal and T. C. Hobson},
+  Year                     = {2017},
+
+  Month                    = {Jan},
+
+  Doi                      = {10.5281/zenodo.260178},
+  Url                      = {https://github.com/neutrons/web_reflectivity}
+}
+
+@Other{Kienzle,
+  Title                    = {Refl1D: Interactive depth profile modeler},
+  Author                   = {Kienzle, PA and Krycka, JA and Patel, N},
+  Url                      = {https://github.com/reflectometry/refl1d}
 }
 
 @Article{Mason2006,
@@ -43,25 +50,24 @@
   Year                     = {2006},
   Pages                    = {955 - 960},
   Volume                   = {385},
+
   Doi                      = {http://dx.doi.org/10.1016/j.physb.2006.05.281},
   ISSN                     = {0921-4526},
   Keywords                 = {Pulsed neutron},
   Url                      = {http://www.sciencedirect.com/science/article/pii/S092145260601177X}
 }
 
-@article{doucet2017,
-  author = {M. Doucet and R. M. Ferraz Leal and T. C. Hobson},
-  title  = {ORNL Reflectivity Fitting Interface},
-  url    = {https://github.com/neutrons/web_reflectivity},
-  DOI={10.5281/zenodo.260178},
-  year={2017}, month={Jan}
+@Online{PyPi,
+  Title                    = {PyPI - the Python Package Index},
+  Author                   = {PyPi},
+  Url                      = {https://pypi.python.org/pypi/django-remote-submission}
 }
 
-
-@Online{Kienzle,
-  title={Refl1D: Interactive depth profile modeler},
-  author={Kienzle, PA and Krycka, JA and Patel, N},
-  url={https://github.com/reflectometry/refl1d}
+@Online{Redis,
[…]
