### 2017-08-15 #366 @ricleal
`paper.md` and `paper.bib` have been updated.
