### 2017-08-16 ornl-ndav/django-remote-submission#33 @djmitche
I see an 0.11.1 tag, but now setup.py says 0.11.2 :)
