### 2017-08-16 fc6e7cb
reorder example instructions
In particular, celery won't start until the example's requirements.txt has been installed.
