### 2017-08-16 #366 @djmitche
That's almost all the checkmarks (just **Version** left)!

This is a really nice piece of software, and *extremely* nicely packaged for re-use.  A few minor details which would make this a little stronger, but do not detract from an "accept" grade:
 * The documentation mixes developer-oriented stuff (how to run tests, how to release) with user-oriented stuff (model, tasks, etc.), and in fact the user-oriented stuff appears last.  A little reorganization there could help a new user to quickly find the information they need and not skim the release process wondering "how does this help me??"
 * The tests don't pass for me.  I expect it's something with my SSH config, and I [filed an issue](https://github.com/ornl-ndav/django-remote-submission/issues/39) about it.  I can see that the test pass in Travis so I trust that they are good tests!
