### 2017-08-16 #366 @djmitche
Oh, and I apologize for the slew of pull requests (5 PRs, 2 issues) -- I take the view that if it's easier for me to just fix something, then that's the appropriate way to provide feedback :)
