### 2017-08-16 #366 @ricleal
Thanks for the feedback. I will address all the issues tomorrow.
