### 2017-08-16 acdc7c8
Better documentation
