### 2017-08-16 80e5b4d
Update HISTORY.rst
