+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 -------
 
+0.11.3 (2017-08-16)
++++++++++++++++++++
+
+* Improved documentation
+
 0.11.2 (2017-08-15)
 +++++++++++++++++++
 
