### 2017-08-16 8f65114
Bump version: 0.11.2 → 0.12.0
