+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.11.2
+current_version = 0.12.0
 commit = True
 tag = True
 
@@ -11,9 +11,10 @@ tag = True
 universal = 1
 
 [flake8]
-exclude =
-  venv,
-  __pycache__,
-  build,
-  dist,
-  migrations
+exclude = 
+	venv,
+	__pycache__,
+	build,
+	dist,
+	migrations
+
