+++ HISTORY.rst
@@ -3,7 +3,7 @@
 History
 -------
 
-0.11.3 (2017-08-16)
+0.12.0 (2017-08-16)
 +++++++++++++++++++
 
 * Improved documentation
