### 2017-08-16 #366 @djmitche
OK, that's everything.  I don't know what happens next (it's my first JOSS review)..
