### 2017-08-16 #366 @ricleal
Great. I've been trying to reproduce your test failures but without success... I will try in a empty docker container later. Also I will try to get the `extras_require` in `setup.py` as you suggested. This was an issue that should have been addressed long time ago...
