### 2017-08-16 #366 @djmitche
To be clear regarding the review, those were just suggestions for further improvement -- this is a strong "accept" for JOSS right now.
