### 2017-08-16 #366 @ricleal
Sure. I just want to make it as user friendly as possible so other people can use it.
I forgot to mention. I also updated/restructured the documentation as suggested.
