### 2017-08-16 0a35996
Added LocalWrapper
