+++ django_remote_submission/wrapper/__init__.py
@@ -1,0 +1,8 @@
+"""
+
+This package contains the wrappers to submit and execute jobs.
+
+``RemoteWrapper`` is the main Wrapper. ``LocalWrapper`` inherits from it but
+runs in the same machine that submitts the jobs.
+
+"""
