+++ docs/modules/wrapper.rst
@@ -1,0 +1,24 @@
+Wrapper
+=======
+
+.. automodule:: django_remote_submission.wrapper
+
+RemoteWrapper
+-------------
+
+.. automodule:: django_remote_submission.wrapper.remote
+
+.. autoclass:: django_remote_submission.wrapper.remote.RemoteWrapper
+   :members:
+
+   .. automethod:: __init__
+   .. automethod:: deploy_key_if_it_does_not_exist
+   .. automethod:: delete_key
+
+LocalWrapper
+-------------
+
+.. automodule:: django_remote_submission.wrapper.local
+
+.. autoclass:: django_remote_submission.wrapper.local.LocalWrapper
+   :members:

[…]
