+++ setup.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     from distutils.core import setup
 
-version = '0.12.0'
+version = '0.13.0'
 
 if sys.argv[-1] == 'publish':
     try:
