### 2017-08-17 f7c3a89
Corrected authors. Changed sphinx theme
