+++ docs/conf.py
@@ -112,8 +112,8 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'django-remote-submission'
-copyright = '2017, Tanner Hobson'
-author = 'Tanner Hobson'
+copyright = '2017, Oak Ridge National Laboratory'
+author = 'T. Hobson, M. Doucet and R. M. Ferraz Leal'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -148,7 +148,7 @@ todo_include_todos = True
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'alabaster'
+html_theme = 'bizstyle'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
