+++ setup.py
@@ -43,13 +43,22 @@ setup(
     ],
     include_package_data=True,
     install_requires=[
-        "django>=1.9.6",
-        "djangorestframework",
-        "django-model-utils>=2.0",
-        "paramiko>=2.0.2",
-        "six>=1.10.0",
-        "channels",
+        # Web framework
+        'django>=1.11.4',
+        # Helpful utilities for creating models in Django
+        'django-model-utils>=3.0.0',
+        # Library to help manage REST APIs
+        'djangorestframework>=3.6.3',
+        # Library to manage connecting to remote hosts via SSH
+        'paramiko>=2.2.1',
+        # Library to help manage python 2/3 compatibility
+        'six>=1.10.0',
+        # Celery
+        'celery>=4.1',
+        # Add websocket connections to Django
+        'channels>=1.1.6',
     ],
+    # install_requires=reqs,
     license="ISCL",
     zip_safe=False,
     keywords='django-remote-submission',
