### 2017-08-17 a6b5723
Removed python 3.4 added Django 1.11
