+++ tox.ini
@@ -1,8 +1,8 @@
 [tox]
 envlist =
-    {py27,py34,py35}-django18
-    {py27,py34,py35}-django19
-    {py27,py34,py35}-django110
+    {py27,py35}-django18
+    {py27,py35}-django19
+    {py27,py35}-django110
 
 [testenv]
 setenv =
@@ -12,5 +12,6 @@ deps =
     pytest
     django-18: Django>=1.8,<1.9
     django-19: Django>=1.9,<1.10
-    django-110: Django>=1.10
+    django-110: Django>=1.10,<1.11
+    django-110: Django>=1.11
     -r{toxinidir}/requirements_test.txt
