### 2017-08-17 818e573
remote is now a flag
