+++ tests/test_tasks.py
@@ -12,13 +12,7 @@ import collections
 import pytest
 import textwrap
 import os
-import logging
 import sys
-from django_remote_submission.wrapper.remote import RemoteWrapper
-from django_remote_submission.wrapper.local import LocalWrapper
-
-
-logging.basicConfig(level=logging.DEBUG, stream=sys.stdout)
 
 
 def pairwise(iterable):
@@ -170,21 +164,21 @@ def job_model_saved(mocker):
     pre_save.disconnect(mock, sender=Job)
 
 
-@pytest.fixture(params=[True, False], ids=["LocalWrapper", "RemoteWrapper"])
-def wrapper_cls(request):
+@pytest.fixture(params=[True, False], ids=["Remote", "Local"])
+def runs_remotely(request):
     '''
-    params == True: return LocalWrapper
-    params == False: return RemoteWrapper
+    params == True: Uses RemoteWrapper
+    params == False: Uses LocalWrapper
     if it's running with command line paramter '--ci' skip it
     '''
 
-    if request.param:
+    if not request.param:
         # return RemoteWrapper
-        return LocalWrapper
+        return False
     elif pytest.config.getoption('--ci'):
         pytest.skip('running on continuous integration')
     else:
-        return RemoteWrapper
+        return True
 
 
 @pytest.mark.django_db
@@ -195,12 +189,12 @@ for i in range(5):
     print("line: {}".format(i))
     time.sleep(0.1)
 ''')
-def test_submit_job_normal_usage(env, job, job_model_saved, wrapper_cls):
+def test_submit_job_normal_usage(env, job, job_model_saved, runs_remotely):
     from django_remote_submission.models import Job, Log
     from django_remote_submission.tasks import submit_job_to_server
     import datetime
 
-    submit_job_to_server(job.pk, env.remote_password, wrapper_cls=wrapper_cls)
+    submit_job_to_server(job.pk, env.remote_password, remote=runs_remotely)
 
     job = Job.objects.get(pk=job.pk)
     assert job.status == Job.STATUS.success
@@ -228,12 +222,12 @@ for i in range(5):
     print("line: {}".format(i), file=sys.stdout if i % 2 == 0 else sys.stderr)
     time.sleep(0.1)
 ''')
-def test_submit_job_multiple_streams(env, job, wrapper_cls):
+def test_submit_job_multiple_streams(env, job, runs_remotely):
     from django_remote_submission.models import Job, Log
     from django_remote_submission.tasks import submit_job_to_server
     import datetime
 
-    submit_job_to_server(job.pk, env.remote_password, wrapper_cls=wrapper_cls)
+    submit_job_to_server(job.pk, env.remote_password, remote=runs_remotely)
 
     assert Log.objects.count() == 5
 
@@ -256,11 +250,11 @@ def test_submit_job_multiple_streams(env, job, wrapper_cls):
 import sys
 sys.exit(1)
 ''')
-def test_submit_job_failure(env, job, wrapper_cls):
+def test_submit_job_failure(env, job, runs_remotely):
     from django_remote_submission.models import Job, Log
     from django_remote_submission.tasks import submit_job_to_server
 
-    submit_job_to_server(job.pk, env.remote_password, wrapper_cls=wrapper_cls)
+    submit_job_to_server(job.pk, env.remote_password, remote=runs_remotely)
 
     job = Job.objects.get(pk=job.pk)
     assert job.status == Job.STATUS.failure
@@ -275,11 +269,11 @@ for i in range(5):
     print('line: {}'.format(i), file=sys.stdout)
     time.sleep(0.1)
 ''')
-def test_submit_job_log_policy_log_total(env, job, wrapper_cls):
+def test_submit_job_log_policy_log_total(env, job, runs_remotely):
     from django_remote_submission.models import Job, Log
[…]
