+++ tests/conftest.py
@@ -1,10 +1,12 @@
 from django.conf import settings
 
+
 def pytest_addoption(parser):
     parser.addoption(
         '--ci', action='store_true',
         help='disable tests that do not work on continuous integration',
     )
+
 
 def pytest_configure():
     settings.configure(
@@ -13,8 +15,8 @@ def pytest_configure():
         DATABASES={
             "default": {
                 "ENGINE": "django.db.backends.sqlite3",
-                "OPTIONS": { # for concurrent writes
-                    "timeout": 30000, # ms
+                "OPTIONS": {  # for concurrent writes
+                    "timeout": 30000,  # ms
                 }
             }
         },
@@ -28,7 +30,7 @@ def pytest_configure():
         SITE_ID=1,
         MIDDLEWARE_CLASSES=(),
         MEDIA_ROOT='media',
-        LOGGING = {
+        LOGGING={
             'version': 1,
             'disable_existing_loggers': False,
             'formatters': {
@@ -58,11 +60,22 @@ def pytest_configure():
                 },
             },
         },
-        CHANNEL_LAYERS = {
+        CHANNEL_LAYERS={
             'default': {
                 "BACKEND": "asgiref.inmemory.ChannelLayer",
                 'ROUTING': 'django_remote_submission.routing.channel_routing',
             },
-        }
-
+        },
+        # Celery configuration
+        BROKER_BACKEND='memory',
+        CELERY_ALWAYS_EAGER=True,
+        CELERY_EAGER_PROPAGATES_EXCEPTIONS=True,
     )
+
+# This is to configure celery: NOT in use
+# import pytest
+# from example.server.celery import app
+# @pytest.fixture(scope='module')
+# def celery_app(request):
+#     app.conf.update(CELERY_ALWAYS_EAGER=True)
+#     return app
