### 2017-08-17 fbc8d2d
Bump version: 0.13.0 → 1.0.0
