+++ HISTORY.rst
@@ -3,9 +3,10 @@
 History
 -------
 
-0.14.0 (2017-08-17)
+1.0.0 (2017-08-17)
 +++++++++++++++++++
 
+* All ready to be released
 * Tasks have an attribute to run locally or remotelly.
 
 0.13.0 (2017-08-17)
