+++ HISTORY.rst
@@ -2,6 +2,10 @@
 
 History
 -------
+1.0.1 (2017-08-18)
++++++++++++++++++++
+
+* Updated DOC with the ``remote`` argument.
 
 1.0.0 (2017-08-17)
 +++++++++++++++++++
