### 2017-08-18 ef22287
Handle the case when the directory does no exist #17
