### 2017-08-18 6610f5a
Swapped print by log
