### 2017-08-18 efa5e67
Bump version: 1.0.1 → 1.1.0
