### 2017-08-18 2214d00
Make Travis happy in Python 2.7
