+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 -------
 
+1.1.1 (2017-08-18)
++++++++++++++++++++
+
+* Fix issue with python 2.7
+
 1.1.0 (2017-08-18)
 +++++++++++++++++++
 
