### 2017-08-18 b5e01d3
Bump version: 1.1.0 → 1.1.1
