+++ setup.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     from distutils.core import setup
 
-version = '1.1.0'
+version = '1.1.1'
 
 if sys.argv[-1] == 'publish':
     try:
