+++ django_remote_submission/__init__.py
@@ -11,6 +11,6 @@ the models to make this function work.
 
 """
 
-__version__ = '1.1.0'
+__version__ = '1.1.1'
 
 default_app_config = 'django_remote_submission.apps.DjangoRemoteSubmissionConfig'
