### 2017-08-21 fbce9f4
Local wrapper uses queues for the script output
