### 2017-08-21 8525a27
Fixed local wrapper log truncate
