+++ HISTORY.rst
@@ -2,6 +2,11 @@
 
 History
 -------
+1.0.2 (2017-08-18)
++++++++++++++++++++
+
+* Fixed Local wrapper truncate the log.
+
 1.0.1 (2017-08-18)
 +++++++++++++++++++
 
