### 2017-08-21 d2e5fe2
Bump version: 1.1.1 → 1.1.2
