### 2017-08-22 97a845c
Local does not truncate in any DB
