### 2017-08-22 adeb333
Bump version: 1.1.2 → 1.1.3
