### 2017-08-22 41bd2da
Fix test running as remote in travis
