+++ HISTORY.rst
@@ -3,6 +3,16 @@
 History
 -------
 
+1.1.5 (2017-08-23)
++++++++++++++++++++
+
+* Local Wrapper does no support Live Log ever.
+
+1.1.4 (2017-08-19)
++++++++++++++++++++
+
+* Fixed CI tests.
+
 1.1.3 (2017-08-19)
 +++++++++++++++++++
 
