### 2017-08-23 79c31d0
Local Wrapper uses process.comunicate
