### 2017-08-23 6390d14
Bump version: 1.1.5 → 1.1.6
