### 2017-08-23 63b4e7c
Update paper.md
