### 2017-08-23 #366 @ricleal
Just a quick heads up:
- I have added Acknowledgments section to the paper (a required statement that I forgot...)
- Since this contribution was accepted shouldn't the ticket be closed?
