### 2017-08-23 a00f6e0
Update README.rst
