### 2017-08-24 d68d912
Fixing up paper structure
