### 2017-08-25 #366 @ricleal
Done. Here it goes:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.848749.svg)](https://doi.org/10.5281/zenodo.848749)

I will update the Zenodo metadata once it gets published.
